+++ src/runtime_src/xdp/profile/device/device_intf.h
@@ -99,7 +99,7 @@ namespace xdp {
     void configAmContext(const std::string& ctx_info);
 
     // Underlying Device APIs
-    uint32_t allocTraceBuf(uint64_t sz ,uint8_t memIdx);
+    size_t allocTraceBuf(uint64_t sz ,uint8_t memIdx);
     void freeTraceBuf(uint32_t bufHandle);
     void* syncTraceBuf(uint32_t bufHandle ,uint64_t offset, uint64_t bytes);
     uint64_t getDeviceAddr(uint32_t bufHandle);
